requirements.txt
# Lightweight requirements for CPU-only, performance-optimized deployment
psutil>=5.9.0
PyPDF2>=3.0.0
python-docx>=0.8.11
click==8.1.7

# Original heavy dependencies (commented out for lightweight deployment)
# nltk==3.8.1
# scikit-learn==1.3.0
# numpy==1.24.3
# pandas==2.0.3
# spacy==3.6.1
# sentence-transformers==2.2.2
# transformers==4.33.2
# torch==2.0.1
